requirements.txt
discord.py>=2.2.2
aiohttp>=3.8.5
python-dotenv>=1.0.1
requests>=2.32.0
